// types/config.ts
export type MonitorTarget = {
  id: string
  name: string
  method: string
  target: string
  enabled?: boolean
  statusPageLink?: string
  expectedCodes?: number[]
  timeout?: number
  headers?: { [key: string]: string | number }
  body?: string
  responseKeyword?: string
  responseForbiddenKeyword?: string
  expectedJson?: { [key: string]: string | number | boolean | null }
  checkProxy?: string
  checkProxyFallback?: boolean
}